adding authorisation with user-roles

+++ todo_app/app.py
@@ -1,3 +1,4 @@
+from operator import truediv
 import os
 from urllib import response
 import requests
@@ -7,11 +8,14 @@ from bson.objectid import ObjectId
 from todo_app.item import Item
 from todo_app.view_model import ViewModel
 from flask import Flask, render_template, request, redirect, session
-from flask_login import LoginManager, login_required, UserMixin, login_user
+from flask_login import LoginManager, login_required, UserMixin, login_user, current_user
 
 todo_status = 'To Do'
 doing_status = 'Doing'
 done_status = 'Done'
+
+#ToDo: change this out to work with a database admin user-role
+writer_user_id = '78470201'
 
 def create_app():
     mongo_db_client = pymongo.MongoClient(os.getenv('DATABASE_CONNECTION_STRING'))
@@ -25,39 +29,45 @@ def create_app():
 
     web_application_client = WebApplicationClient(os.getenv('CLIENT_ID'))
 
+    class User(UserMixin):
+        def __init__(self, id):
+            self.id = id
+            if id == writer_user_id:
+                self.user_role = 'writer'
+            else:
+                self.user_role = 'reader'
+
     @login_manager.unauthorized_handler
     def unauthenticated():
         return redirect(web_application_client.prepare_request_uri('https://github.com/login/oauth/authorize'))
     
-    class User(UserMixin):
-        def __init__(self, id):
-            self.id = id
-
     @login_manager.user_loader     
     def load_user(user_id):         
         user = User(user_id)  
+        return user
     
-    login_manager.init_app(app) 
+    login_manager.init_app(app)
 
     @app.route('/login/callback', methods=['GET'])
     def login_user_callback():    
-        data = web_application_client.prepare_request_body(
-            code = request.args['code'],
-            redirect_uri = request.url,
-            client_id = os.getenv('CLIENT_ID'),
-            client_secret = os.getenv('CLIENT_SECRET')
+        token_url, headers, data = web_application_client.prepare_token_request(
+            'https://github.com/login/oauth/access_token', code = request.args['code'], 
+            client_id = os.getenv('CLIENT_ID'), client_secret = os.getenv('CLIENT_SECRET')
         )
-
-        response = requests.post('https://github.com/login/oauth/access_token', data=data)
+        
+        headers['Accept'] = 'application/json'
+        response = requests.post(token_url, headers=headers, data=data)
         web_application_client.parse_request_body_response(response.text)
         access_token = web_application_client.token['access_token']
         header = {'Authorization': f'Bearer {access_token}'}
         response = requests.get('https://api.github.com/user', headers=header).json()
-
         user_id = response['id']
         user = User(user_id)
         login_user(user)
         return redirect('/')
+
+    def is_writer():
+        return current_user.user_role == 'writer'
 
     @app.route('/')
     @login_required
@@ -67,10 +77,13 @@ def create_app():
         for item in items:
             cards.append(Item(item['_id'], item['name'], item['description'], item['due_date'], item['status']))
         item_view_model = ViewModel(cards)
-        return render_template('index.html', view_model = item_view_model)
+        return render_template('index.html', view_model = item_view_model, is_writer = is_writer())
 
     @app.route('/create-todo/', methods=['POST'])
     def create_todo():
+        if not is_writer():
+            return redirect('/')
+        
         title = request.form.get('title')
         desc = request.form.get('description')
         due = request.form.get('due-date')
@@ -85,6 +98,9 @@ def create_app():
 
     @app.route('/todo/<id>', methods=['POST'])
     def to_do(id):
+        if not is_writer():
+            return redirect('/')
+
         query = {'_id': ObjectId(id)}
         update_values = {'$set': {'status':todo_status}}
         collection.update_one(query, update_values)
@@ -92,6 +108,9 @@ def create_app():
 
     @app.route('/doing/<id>', methods=['POST'])
     def doing(id):
+        if not is_writer():
+            return redirect('/')
+
         query = {'_id': ObjectId(id)}
         update_values = {'$set': {'status':doing_status}}
         collection.update_one(query, update_values)
@@ -99,6 +118,9 @@ def create_app():
 
     @app.route('/done/<id>', methods=['POST'])
     def done(id):
+        if not is_writer():
+            return redirect('/')
+
         query = {'_id': ObjectId(id)}
         update_values = {'$set': {'status':done_status}}
         collection.update_one(query, update_values)
@@ -106,6 +128,9 @@ def create_app():
 
     @app.route('/delete/<id>', methods=['POST'])
     def delete(id):
+        if not is_writer():
+            return redirect('/')
+
         query = {'_id': ObjectId(id)}
         collection.delete_one(query)
         return redirect('/')
